# docs/swagger.yaml
definitions:
...
  main.Paginate-main_Room:
    properties:
      current_page:
        type: integer
      items:
        items:
          $ref: '#/definitions/main.Room'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_page:
        type: integer
    type: object
...
  main.Response-any:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  main.Response-main_Paginate-main_Room:
    properties:
      data:
        $ref: '#/definitions/main.Paginate-main_Room'
      message:
        type: string
    type: object
...
  main.Response-main_Room:
    properties:
      data:
        $ref: '#/definitions/main.Room'
      message:
        type: string
    type: object
  main.Response-main_Seat:
    properties:
      data:
        $ref: '#/definitions/main.Seat'
      message:
        type: string
    type: object
...
  main.Room:
    properties:
      capacity:
        description: relation
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  main.RoomFilter:
    properties:
      ids:
        items:
          type: integer
        type: array
      is_usable:
        type: boolean
      names:
        items:
          type: string
        type: array
    type: object
  main.RoomInput:
    properties:
      name:
        type: string
    type: object
  main.Seat:
    properties:
      additional_price:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_available:
        description: relation
        type: boolean
      name:
        type: string
      room_id:
        type: integer
      updated_at:
        type: string
    type: object
  main.SeatInput:
    properties:
      additional_price:
        type: integer
      name:
        type: string
      room_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/rooms:
    post:
      consumes:
      - application/json
      description: admin create room
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.RoomInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Create Room
      tags:
      - rooms
  /api/admin/rooms/{id}:
    delete:
      consumes:
      - application/json
      description: admin delete room by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Delete Room
      tags:
      - rooms
    put:
      consumes:
      - application/json
      description: admin update room by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.RoomInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Update Room
      tags:
      - rooms
  /api/admin/rooms/{id}/seats:
    post:
      consumes:
      - application/json
      description: admin set seat for room by room id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: body request
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/main.SeatInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Set room seats
      tags:
      - rooms
...
  /api/rooms/{id}:
    get:
      consumes:
      - application/json
      description: get room by id
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Get Room
      tags:
      - rooms
  /api/rooms/{id}/seats:
    get:
      consumes:
      - application/json
      description: get seats by room id
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Seat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Get room seats
      tags:
      - rooms
  /api/rooms/filter:
    post:
      consumes:
      - application/json
      description: filter rooms
      parameters:
      - description: pagination page
        in: query
        name: page
        type: integer
      - description: pagination page size
        in: query
        name: per_page
        type: integer
      - description: filter
        in: body
        name: request
        schema:
          $ref: '#/definitions/main.RoomFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Paginate-main_Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Filter Room
      tags:
      - rooms
...
swagger: "2.0"